# music_app_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.UpdateUserRequest:
    properties:
      display_name:
        maxLength: 50
        minLength: 2
        type: string
      photo_url:
        type: string
      preferences:
        additionalProperties: true
        type: object
    type: object
  http.UserResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      photo_url:
        type: string
      preferences:
        additionalProperties: true
        type: object
      roles:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      success:
        type: boolean
    type: object
...
host: localhost:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the Music App, providing endpoints for
    authentication, music discovery, playlists, and user library management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Music App Backend API
  version: 1.2.0
paths:
...
  /users/me:
    get:
      description: Retrieves the full profile for the user associated with the JWT
        token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates the display name, photo URL, or preferences for the current
        user.
      parameters:
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      security:
      - Bearer: []
      summary: Update current user profile
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"